Refactor mathematical functions into an explicit feature
